fix(orchestrator): validation order with ui:order
