+++ workspaces/orchestrator/.changeset/twenty-fishes-count.md
@@ -1,0 +1,5 @@
+---
+'@red-hat-developer-hub/backstage-plugin-orchestrator-form-react': patch
+---
+
+Fixing validation of multi-step wizard when ui:order is used by supplying correct step fields.
